improve startup procedure

+++ app/lib/grape.js
@@ -8,6 +8,7 @@ var util = require('util');
 var cluster = require('cluster');
 var g_app = require(__dirname + '/app.js');
 var comms = require(__dirname + '/comms.js');
+var async = require('async');
 
 exports = module.exports = function(_o) {
 	this.self = this;
@@ -19,16 +20,50 @@ exports = module.exports = function(_o) {
 		{
 			var pidfile = this.options.log_directory + '/grape.pid';
 
-			var channel = new comms.server(_o);
-			channel.on('error', function(message) {
-				//TODO
-			});
-			channel.start();
+			// check if pidfile exists, if it does kill the process and delte the file
+			var start_pidfile = function(next) {
+				console.log("Setting up PID file " + pidfile + " ...");
+				if (fs.existsSync(pidfile))
+				{
+					console.log("PID file exists");
+					//check if process exists
+					var old_pid = fs.readFileSync(pidfile, 'UTF8');
+					var proc_running = false;
+					try {
+						proc_running = process.kill(old_pid, 0);
+					} catch (e) {
+						proc_running = false;
+					}
+					if (proc_running)
+					{
+						console.log("Killing running process " + old_pid + " (found from pidfile)");
+						process.kill(old_pid, 'SIGINT');
+						setTimeout(function() { 
+							fs.writeFileSync(pidfile, process.pid.toString());
+							next(); 
+						}, 2000);
+					}
+					else
+					{
+						console.log("Overwriting PID file ");
+						fs.writeFileSync(pidfile, process.pid.toString());
+						next();
+					}
+				}
+				else
+				{
+					console.log("PID does not exist");
+					fs.writeFileSync(pidfile, process.pid.toString());
+					next();
+				}
 
-			var start_instances = function() {
-				fs.writeFileSync(pidfile, process.pid.toString());
+			}
+
+			//start worker instances
+			var start_instances = function(next) {
 
 				process.on('exit', function(code) {
+					console.log("Process exiting ");
 					fs.unlinkSync(pidfile);
 				});
 				process.on('SIGINT', function(code) {
@@ -46,33 +81,21 @@ exports = module.exports = function(_o) {
 				{
 					self.createWorker();
 				}
+
+				next();
 			};
 
-			if (fs.existsSync(pidfile))
-			{
-				//check if process exists
-				var old_pid = fs.readFileSync(pidfile, 'UTF8');
-				var proc_running = false;
-				try {
-					proc_running = process.kill(old_pid, 0);
-				} catch (e) {
-					proc_running = false;
-				}
-				if (proc_running)
-				{
-					console.log("Killing running process " + old_pid + " (found from pidfile)");
-					process.kill(old_pid, 'SIGINT');
-					setTimeout(start_instances, 2000);
-				}
-				else
-				{
-					start_instances();
-				}
-			}
-			else
-			{
-				start_instances();
-			}
+
+			var start_comms_channel = function(next) {
+				var channel = new comms.server(_o);
+				channel.on('error', function(message) {
+					//TODO
+				});
+				channel.start();
+				next();
+			};
+
+			async.series([start_pidfile, start_comms_channel, start_instances]);
 		}
 		else
 		{
